Build succeeded! Still seeing an localdef error though...
