+++ FIRMWARE/PROJECT_CODE/LCCM655__RLOOP__FCU_CORE/FLIGHT_CONTROLLER/MAIN_SM/pod__types.h
@@ -3,7 +3,6 @@
 
 // #include "../COMPAT.h"
 #include "../../state_machine.h"
-#include "../../timers.h"
 
 
 /*  Now in fcu_core__types.h
